requirements.txt
pyarrow==18.1.0
kafka-python
great-expectations
confluent-kafka
seaborn

# Feature store and model registry
hopsworks

# Be more efficient when making REST (Http) requests
requests-cache==1.2.0
retry-requests==2.0.0

# ML framework libraries
xgboost==2.0.3
scikit-learn==1.4.1.post1

# Plot charts
matplotlib==3.8.3

# API for the tweets from the Bsky app
atproto
graphviz
hyperopt

# Set up the server
flask
flask-cors
gradio